refactor: inline init handlers into initializer
