+++ lib/index.js
@@ -10259,10 +10259,14 @@ var KairoInitEventId = /* @__PURE__ */ ((KairoInitEventId2) => {
 
 // src/router/init/KairoInitializer.ts
 var KairoInitializer = class {
-  constructor(initListener, discoveryHandler, registrationHandler, contextMutator) {
+  constructor(initListener, discoveryManager, discoveryResponder, registrationManager, registrationResponder, context, runtime2, contextMutator) {
     this.initListener = initListener;
-    this.discoveryHandler = discoveryHandler;
-    this.registrationHandler = registrationHandler;
+    this.discoveryManager = discoveryManager;
+    this.discoveryResponder = discoveryResponder;
+    this.registrationManager = registrationManager;
+    this.registrationResponder = registrationResponder;
+    this.context = context;
+    this.runtime = runtime2;
     this.contextMutator = contextMutator;
   }
   subscription;
@@ -10278,28 +10282,33 @@ var KairoInitializer = class {
     this.subscription = void 0;
   }
   handleDiscoveryQuery = (message) => {
-    const kairoId = this.discoveryHandler.handle(message);
+    const kairoId = this.discoveryManager.resolveKairoId(message, this.runtime.currentTick());
+    this.discoveryResponder.respond(kairoId);
     this.contextMutator.setKairoId(kairoId);
   };
   handleRegistrationRequest = (message) => {
-    const registry = this.registrationHandler.handle(message);
+    const registry = this.registrationManager.resolveRegistry(
+      message,
+      this.runtime.currentTick(),
+      this.context.kairoId,
+      this.context.addonProperties
+    );
     if (!registry) return;
+    this.registrationResponder.respond(registry);
     this.contextMutator.setKairoRegistry(registry);
   };
 };
 
 // src/router/factories/KairoInitializerFactory.ts
 var KairoInitializerFactory = class {
-  constructor(runtime2, idRegistryFactory2, listenerFactory2, discoveryFactory2, discoveryResponderFactory2, discoveryHandlerFactory2, registrationFactory2, registrationResponderFactory2, registrationHandlerFactory2) {
+  constructor(runtime2, idRegistryFactory2, listenerFactory2, discoveryFactory2, discoveryResponderFactory2, registrationFactory2, registrationResponderFactory2) {
     this.runtime = runtime2;
     this.idRegistryFactory = idRegistryFactory2;
     this.listenerFactory = listenerFactory2;
     this.discoveryFactory = discoveryFactory2;
     this.discoveryResponderFactory = discoveryResponderFactory2;
-    this.discoveryHandlerFactory = discoveryHandlerFactory2;
     this.registrationFactory = registrationFactory2;
     this.registrationResponderFactory = registrationResponderFactory2;
-    this.registrationHandlerFactory = registrationHandlerFactory2;
   }
   create(context, mutator) {
     const listener = this.listenerFactory.create(this.runtime);
@@ -10309,13 +10318,12 @@ var KairoInitializerFactory = class {
     const registrationResponder = this.registrationResponderFactory.create(this.runtime);
     return new KairoInitializer(
       listener,
-      this.discoveryHandlerFactory.create(discovery, discoveryResponder, this.runtime),
-      this.registrationHandlerFactory.create(
-        registration,
-        registrationResponder,
-        context,
-        this.runtime
-      ),
+      discovery,
+      discoveryResponder,
+      registration,
+      registrationResponder,
+      context,
+      this.runtime,
       mutator
     );
   }
@@ -13127,29 +13135,6 @@ var AddonDiscoveryManagerFactory = class {
   }
 };
 
-// src/router/init/discovery/DiscoveryQueryHandler.ts
-var DiscoveryQueryHandler = class {
-  constructor(discoveryManager, discoveryResponder, runtime2) {
-    this.discoveryManager = discoveryManager;
-    this.discoveryResponder = discoveryResponder;
-    this.runtime = runtime2;
-  }
-  handle = (msg) => {
-    const kairoId = this.discoveryManager.resolveKairoId(msg, this.runtime.currentTick());
-    this.discoveryResponder.respond(kairoId);
-    return kairoId;
-  };
-};
-
-// src/router/init/discovery/factories/DiscoveryQueryHandlerFactory.ts
-var DiscoveryQueryHandlerFactory = class {
-  constructor() {
-  }
-  create(discovery, responder, runtime2) {
-    return new DiscoveryQueryHandler(discovery, responder, runtime2);
-  }
-};
-
 // src/router/init/discovery/response/errors.ts
 var DiscoveryResponseError = class extends Error {
   reason;
@@ -13341,36 +13326,6 @@ var AddonRegistrationManagerFactory = class {
   }
 };
 
-// src/router/init/registration/RegistrationRequestHandler.ts
-var RegistrationRequestHandler = class {
-  constructor(registrationManager, registrationResponder, context, runtime2) {
-    this.registrationManager = registrationManager;
-    this.registrationResponder = registrationResponder;
-    this.context = context;
-    this.runtime = runtime2;
-  }
-  handle = (msg) => {
-    const registry = this.registrationManager.resolveRegistry(
-      msg,
-      this.runtime.currentTick(),
-      this.context.kairoId,
-      this.context.addonProperties
-    );
-    if (!registry) return;
-    this.registrationResponder.respond(registry);
-    return registry;
-  };
-};
-
-// src/router/init/registration/factories/RegistrationRequestHandlerFactory.ts
-var RegistrationRequestHandlerFactory = class {
-  constructor() {
-  }
-  create(registration, responder, context, runtime2) {
-    return new RegistrationRequestHandler(registration, responder, context, runtime2);
-  }
-};
-
 // src/router/init/registration/response/errors.ts
 var RegistrationResponseError = class extends Error {
   reason;
@@ -13557,20 +13512,16 @@ var idRegistryFactory = new ScoreboardIdRegistryFactory();
 var listenerFactory = new KairoInitListenerFactory();
 var discoveryFactory = new AddonDiscoveryManagerFactory();
 var discoveryResponderFactory = new DiscoveryResponderFactory();
-var discoveryHandlerFactory = new DiscoveryQueryHandlerFactory();
 var registrationFactory = new AddonRegistrationManagerFactory();
 var registrationResponderFactory = new RegistrationResponderFactory();
-var registrationHandlerFactory = new RegistrationRequestHandlerFactory();
 var initializerFactory = new KairoInitializerFactory(
   runtime,
   idRegistryFactory,
   listenerFactory,
   discoveryFactory,
   discoveryResponderFactory,
-  discoveryHandlerFactory,
   registrationFactory,
-  registrationResponderFactory,
-  registrationHandlerFactory
+  registrationResponderFactory
 );
 var router = new KairoRouter(
   (context, mutator) => initializerFactory.create(context, mutator)
