+++ src/index.ts
@@ -3,10 +3,8 @@ import { ScoreboardIdRegistryFactory } from "./minecraft/ScoreboardIdRegistryFac
 import { KairoInitializerFactory } from "./router/factories/KairoInitializerFactory";
 import { KairoInitListenerFactory } from "./router/factories/KairoInitListenerFactory";
 import { AddonDiscoveryManagerFactory } from "./router/init/discovery/factories/AddonDiscoveryManagerFactory";
-import { DiscoveryQueryHandlerFactory } from "./router/init/discovery/factories/DiscoveryQueryHandlerFactory";
 import { DiscoveryResponderFactory } from "./router/init/discovery/factories/DiscoveryResponderFactory";
 import { AddonRegistrationManagerFactory } from "./router/init/registration/factories/AddonRegistrationManagerFactory";
-import { RegistrationRequestHandlerFactory } from "./router/init/registration/factories/RegistrationRequestHandlerFactory";
 import { RegistrationResponderFactory } from "./router/init/registration/factories/RegistrationResponderFactory";
 import { KairoRouter } from "./router/KairoRouter";
 
@@ -17,10 +15,8 @@ const idRegistryFactory = new ScoreboardIdRegistryFactory();
 const listenerFactory = new KairoInitListenerFactory();
 const discoveryFactory = new AddonDiscoveryManagerFactory();
 const discoveryResponderFactory = new DiscoveryResponderFactory();
-const discoveryHandlerFactory = new DiscoveryQueryHandlerFactory();
 const registrationFactory = new AddonRegistrationManagerFactory();
 const registrationResponderFactory = new RegistrationResponderFactory();
-const registrationHandlerFactory = new RegistrationRequestHandlerFactory();
 
 const initializerFactory = new KairoInitializerFactory(
     runtime,
@@ -28,10 +24,8 @@ const initializerFactory = new KairoInitializerFactory(
     listenerFactory,
     discoveryFactory,
     discoveryResponderFactory,
-    discoveryHandlerFactory,
     registrationFactory,
     registrationResponderFactory,
-    registrationHandlerFactory,
 );
 
 export const router = new KairoRouter((context, mutator) =>
