+++ src/router/factories/KairoInitializerFactory.ts
@@ -1,9 +1,7 @@
 import { AddonDiscoveryManagerFactory } from "../init/discovery/factories/AddonDiscoveryManagerFactory";
-import { DiscoveryQueryHandlerFactory } from "../init/discovery/factories/DiscoveryQueryHandlerFactory";
 import { DiscoveryResponderFactory } from "../init/discovery/factories/DiscoveryResponderFactory";
 import { KairoInitializer } from "../init/KairoInitializer";
 import { AddonRegistrationManagerFactory } from "../init/registration/factories/AddonRegistrationManagerFactory";
-import { RegistrationRequestHandlerFactory } from "../init/registration/factories/RegistrationRequestHandlerFactory";
 import { RegistrationResponderFactory } from "../init/registration/factories/RegistrationResponderFactory";
 import { KairoContext, KairoContextMutator } from "../KairoContext";
 import { IdRegistryFactory } from "../types/IdRegistryFactory";
@@ -17,10 +15,8 @@ export class KairoInitializerFactory {
         private readonly listenerFactory: KairoInitListenerFactory,
         private readonly discoveryFactory: AddonDiscoveryManagerFactory,
         private readonly discoveryResponderFactory: DiscoveryResponderFactory,
-        private readonly discoveryHandlerFactory: DiscoveryQueryHandlerFactory,
         private readonly registrationFactory: AddonRegistrationManagerFactory,
         private readonly registrationResponderFactory: RegistrationResponderFactory,
-        private readonly registrationHandlerFactory: RegistrationRequestHandlerFactory,
     ) {}
 
     create(context: KairoContext, mutator: KairoContextMutator): KairoInitializer {
@@ -34,13 +30,12 @@ export class KairoInitializerFactory {
 
         return new KairoInitializer(
             listener,
-            this.discoveryHandlerFactory.create(discovery, discoveryResponder, this.runtime),
-            this.registrationHandlerFactory.create(
-                registration,
-                registrationResponder,
-                context,
-                this.runtime,
-            ),
+            discovery,
+            discoveryResponder,
+            registration,
+            registrationResponder,
+            context,
+            this.runtime,
             mutator,
         );
     }
