+++ src/router/init/KairoInitializer.ts
@@ -1,9 +1,12 @@
 import { Disposable } from "../types/Disposable";
-import { KairoContextMutator } from "../KairoContext";
-import { DiscoveryQueryHandler } from "./discovery/DiscoveryQueryHandler";
+import { KairoContext, KairoContextMutator } from "../KairoContext";
+import { KairoRuntime } from "../types/KairoRuntime";
 import { KairoInitEventId } from "./KairoInitEventId";
 import { KairoInitListener } from "./KairoInitListener";
-import { RegistrationRequestHandler } from "./registration/RegistrationRequestHandler";
+import { AddonDiscoveryManager } from "./discovery/AddonDiscoveryManager";
+import { DiscoveryResponder } from "./discovery/DiscoveryResponder";
+import { AddonRegistrationManager } from "./registration/AddonRegistrationManager";
+import { RegistrationResponder } from "./registration/RegistrationResponder";
 
 // kjs-router-ch 0010
 export class KairoInitializer implements Disposable {
@@ -11,8 +14,12 @@ export class KairoInitializer implements Disposable {
 
     constructor(
         private readonly initListener: KairoInitListener,
-        private readonly discoveryHandler: DiscoveryQueryHandler,
-        private readonly registrationHandler: RegistrationRequestHandler,
+        private readonly discoveryManager: AddonDiscoveryManager,
+        private readonly discoveryResponder: DiscoveryResponder,
+        private readonly registrationManager: AddonRegistrationManager,
+        private readonly registrationResponder: RegistrationResponder,
+        private readonly context: KairoContext,
+        private readonly runtime: KairoRuntime,
         private readonly contextMutator: KairoContextMutator,
     ) {}
 
@@ -31,14 +38,23 @@ export class KairoInitializer implements Disposable {
     }
 
     private handleDiscoveryQuery = (message: string): void => {
-        const kairoId = this.discoveryHandler.handle(message);
+        const kairoId = this.discoveryManager.resolveKairoId(message, this.runtime.currentTick());
+
+        this.discoveryResponder.respond(kairoId);
         this.contextMutator.setKairoId(kairoId);
     };
 
     private handleRegistrationRequest = (message: string): void => {
-        const registry = this.registrationHandler.handle(message);
+        const registry = this.registrationManager.resolveRegistry(
+            message,
+            this.runtime.currentTick(),
+            this.context.kairoId,
+            this.context.addonProperties,
+        );
+
         if (!registry) return;
 
+        this.registrationResponder.respond(registry);
         this.contextMutator.setKairoRegistry(registry);
     };
 }
